Android.mk: Stop using version.sh script
This isn't producing the correct results anyways, since the working
directory of the script is the top level of the android tree, not
external/iw like the script expects. And moving the working directory
doesn't help either, since we don't have a copy the git tags that it
expects.

I noticed this while testing git submodules as an alternative to repo,
where we suddenly have a .git directory at the top level, and the script
failed.

So just extract the version number itself from the script and write out
our own version.c file.

Test: m out/target/product/generic/gen/EXECUTABLES/iw_intermediates/version.c
Test: Compare version.c before/after
Change-Id: I98ff4e44b84afbcd2ce8319ccf6816d5fa57a90a